OPERATIONS WORKING BUT WITH MISTAKES
The calculator can add, substract, multiply and divide, but doesn't imitate a real calculator properly.
The first aim to achive is change it's behavior, like a real calculator.

Besides, we need to implement a filter of the result of the calculator, and limitate the number of digits of the output