DO OVERFLOW ERROR AND DECIMAL FILTER
If the result is a extremely big (or negative big) number, the calculator shows a error message and blocks itself
For the decimal results, we calculate it with the toPrecision() method. Anyway, is possible that a result with more
digits than the MAX_NUM_OF_DIGITS or in scientific notation could be shown on screen.
As it does not affect to the prompt, we have decided not to perfection these screen displays, because we consider they
would be out of the scope of this exercise.